requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.25.0
anyio==3.7.1
apify_client==1.9.2
apify_shared==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.30.1
azure-storage-blob==12.19.0
backoff==2.2.1
bidict==0.23.1
blinker==1.6.2
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
cloudevents==1.9.0
colorama==0.4.6
contourpy==1.3.1
crashtest==0.4.1
cryptography==42.0.7
cycler==0.12.1
Cython==0.29.32
dataclasses-json==0.6.3
decorator==5.2.1
deprecation==2.1.0
distlib==0.3.7
distro==1.9.0
dnspython==2.4.2
dulwich==0.21.6
email-validator==2.1.0.post1
eval_type_backport==0.2.0
eventlet==0.34.2
faiss-cpu==1.7.4
fastapi==0.104.1
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
fonttools==4.56.0
frozenlist==1.4.0
fsspec==2023.10.0
functions-framework==3.4.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.41
google_search_results==2.4.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.1
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
keyring==24.2.0
kiwisolver==1.4.8
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langsmith==0.0.83
loguru==0.7.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.10.1
mdurl==0.1.2
minio==7.2.3
more-itertools==10.1.0
motor==3.3.2
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.1
openai==0.27.9
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
pandasql==0.7.3
pexpect==4.8.0
pillow==10.4.0
pinecone-client==3.0.1
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.5.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
proglog==0.1.10
propcache==0.2.0
psutil==7.0.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyhumps==3.8.0
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.2.1
pypdf==4.0.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==2.15.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.3
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
textbase-client==0.1.3
threadpoolctl==3.6.0
tiktoken==0.5.2
together==1.3.3
tokenizers==0.20.1
tomlkit==0.12.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.5
transformers==4.45.2
trove-classifiers==2023.8.7
typer==0.12.5
typing-inspect==0.9.0
typing_extensions
tzdata==2023.4
urllib3==2.0.4
uvicorn==0.24.0.post1
validate_email==1.3
validators==0.28.1
virtualenv==20.24.4
watchdog==3.0.0
Werkzeug==2.3.7
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.15.5
yaspin==3.0.0
zipp==3.16.2
